% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_roubo_veiculos_ssp_2020_2021.R
\docType{data}
\name{dados_roubo_veiculos_ssp_2020_2021}
\alias{dados_roubo_veiculos_ssp_2020_2021}
\title{Dados sobre o roubo de 2974 veiculos em 2020 e 2021}
\format{
Um data.frame com 2974 linhas e 32 colunas
\describe{
\item{num_bo}{numero identificador do boletim de ocorrência}
\item{ano_bo}{ano de registro do boletim de ocorrência}
...
}
}
\source{
\url{http://www.ssp.sp.gov.br/transparenciassp/Consulta.aspx}
}
\usage{
dados_roubo_veiculos_ssp_2020_2021
}
\description{
Um conjunto de dados contendo informações sobre veículos
roubados no estado de Sao Paulo em 2020 e 2021.
}
\keyword{datasets}
